Add function to create Aptfile.tmp
This function saves all packagaes which have been installed via 'apt* install'
AND which are NOT already in the real Aptfile in a file called Aptfile.tmp.

Now I have anachron run this once a day and the tmp file will show me which
packages I might want to add to the Aptfile over time.