Add wincent/corpus to plugins (finally)
I had to update to neovim 5.0 for this to work. Fortunately using the
appimage format is quite easy. All I have to do now is to change the
references to the old vim.

Also learned about -s <scriptin> which allows me to start Corpus from
tmux in a new pane from anywhere. (vi +Corpus did not work cause of some
lua loading issue)

I wanted to set this up for a year now. Happy I finally got around to
it. Noice!